Relax T: std::cmp::Ord on CKMS
The use of BinaryHeap internal to CKMS forced T to be constraint
to types that implemented std::cmp::Ord. However, given that the
internal buffer is small-ish outside of conditions where CKMS
error is very, very small we prefer to do a sorted insert into
a Vec. As noted, the present insertion is O(n) and this is a
clear place for optimization.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>