[CKMS] Insert into samples immediately
The CKMS paper discusses buffering points for batch
or 'cursor' insertion. After some experimentation it
was determined that this buffering decreased
throughput in the present implementation. A benchmark
to stress the machine to its limits is desirable.

On the plus side, there is no longer any need for a
flush_0 function and the user will not be tripped up by
a lack of points that have been inserted but not removed
from buffer.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>