Make CKMS serializable, impl AddAssign
This commit fiddles with the interface of CKMS a bit by adding
the AddAssign trait--useful for merging two CKMS's together or
for using CKMS in a numeric context--and serde serialization. It's
now possible to flush this happy thing to disk or over the wire.

A few bugs were corrected:

  * It's now possible to set the error bounds, which previously
    was hard-coded at 0.001 and
  * partial_cmp is used in place of home-grown comparision.

This marks version 0.2.0 of the quantiles library.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>