Fix the minimum CKMS error bound at 0.000_000_000_1
This commit fixes the minimum error bound at one in one-billion to
avoid a truncation to zero issue when we convert the insert_threshold
from a float to usize. This is kind of a lame fix--it'd be better to
take into account the value of std::usize::MAX--or avoid the float ->
usize conversion step entirely.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>