First cut of histogram type in quantiles
This commit introduces a binning histogram into quantiles. The
implementation does not require equally sized bins and is generic
with regard to integer / floating types. This makes the implementation
a little weird but, all in all, it should be useful if not the fastest
kid on the block.

There are improvements that need to be made. Namely we lack an Iterator
implementation and the documentation is bare bones. This commit is
intended for review and comment by interested parties.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>