First cut of histogram type in quantiles (#16)
* First cut of histogram type in quantiles

This commit introduces a binning histogram into quantiles. The
implementation does not require equally sized bins and is generic
with regard to integer / floating types. This makes the implementation
a little weird but, all in all, it should be useful if not the fastest
kid on the block.

There are improvements that need to be made. Namely we lack an Iterator
implementation and the documentation is bare bones. This commit is
intended for review and comment by interested parties.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add documentation to Histogram implementation

This commit brings in documentation for the Histogram implementation.
Need a little review on the text, I imagine, since there's lots of
ways to make a Histogram and lots of histograms to cloud the mind.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce summation of Histogram inputs

Thsi commit properly sums the inputs of Histogram across all
bins. It would be straightforward to keep statistics in the bins
themselves, basically anything that can be computed online.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce Iterator implementation for Histogram

This commit introduces an Iterator for the histogram over the bins.
The expectation is that a consumer which needs to view all the bins
will iterate them, rather than storing the configured bins.

I've updated the tests some to cover all the non-nightly types.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Update documentation per Tom Santero's request

I've added notes on the T of Histogram and concerns around it in
overflow situations. There's also been a cargo fmt run here, which
has bounced some things around.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>