Histogram cernan support (#17)
* Implement support for cernan in Histogram

This commit adds Serialize / Deserialize to the Histogram derivation
and also adds on PartialEq. These are things cernan needs.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Add conversion of Histogram into tuples

Cernan needs this function so that it can convert Histograms
into a thing to send over the wire via its native protocol. Bit
goofy maybe.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Ensure we AddAssign histogram sums

This commit ensures that the sums of histograms get added together
properly when we AddAssign two histograms. This was overlooked.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>