Cut 0.6.0
This release introduces the Histogram type, a binning summarization
of unequal, explicit bins. The primary use-case for this is in the
postmates/cernan project for the BIN aggregation method but it is
hoped that others will find it useful. The implementation is biased
toward fast writes at the expense of read speed, though not overmuch.

There are crash possibilities owing to the lack of generic checked
addition. Please read the documentation around Histogram for more
details.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>