Check in selist
This is a naive implementation of an unrolled linked list, plus
benchmarks. It's not... great, compared to a Vec. I'm not sure if
we can improve on this. plflib's benchmarks[1] indicate that list
is not much better than vector, in terms of insertion performance
which is our main concern here.

[1] http://www.plflib.org/benchmarks_haswell_gcc.htm#rawperformance

Signed-off-by: Brian L. Troutwine <blt@postmates.com>