Final wrap-up
This has been a long chain of commits. There's interesting work in
here -- I think I could probably contribute SEList to intrusive
collections -- but nothing that scooted up the performance for Cernan.
The prometheus sink in cernan would need to have some kind of quantile
summary that doesn't keep a sub-linear but growing number of points.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>