Bump CKMS insertion performance (#19)
* Better benchmarks

It turns out that CKMS gets very slow once you start adding
more and more points in. That is... bad. Real bad.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Modestly improve insert performance

This commit removes the computation of a variable behind a branch
gate, with modest improvement to priv_insert. The name of the game
here will be these micro-optimizations until I can find some major
smoking gun.

    name                    control ns/iter  variable ns/iter  diff ns/iter   diff %  speedup
    u16::bench_insert_100   5,973            5,014                     -959  -16.06%   x 1.19
    u16::bench_insert_1000  324,281          212,675               -111,606  -34.42%   x 1.52

Control is fc1dbb4aae52a4c7d9555694424b8806d213bfa3 until otherwise
noted.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Base on new benchmarks

I need much more of a sample all the way through u16. This commit
is that.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Modestly improve 'invariant'

    name                     control ns/iter  variable ns/iter  diff ns/iter  diff %  speedup
    u16::bench_insert_100    4,780            4,657                     -123  -2.57%   x 1.03
    u16::bench_insert_1000   192,389          189,745                 -2,644  -1.37%   x 1.01
    u16::bench_insert_10000  11,649,469       11,047,675            -601,794  -5.17%   x 1.05

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Freebie

This commit does not hurt or improve performance but it does skip
a conversion.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce intrinsic-collections LinkedList

This commit blows the doors off our performance issue with many
points, as you'd expect from a linked list. There's a fair bit of
code commented out and I bet we broke serde but we'll find out.

    name                     control ns/iter  variable ns/iter  diff ns/iter   diff %  speedup
    u16::bench_insert_100    4,639            3,196                   -1,443  -31.11%   x 1.45
    u16::bench_insert_1000   190,038          51,656                -138,382  -72.82%   x 3.68
    u16::bench_insert_10000  11,108,320       536,707            -10,571,613  -95.17%  x 20.70

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Now passing extant tests

We're still down on the serde serialization -- which we will have to
do by hand -- and some of the top-level linters. Progress though.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Manually implement serde serialization

Boy, it turns out this is not easy to do. The LinkedList has no
serde serialization and that's okay because we can do it ourselves.
I don't have high confidence in this code yet on account of there's
no unit tests but it does at least compile.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Revert introduction of linked list

While promising the overhead of the linked list is not feasible.
The constant operations kill us, especially since compression means
that the actual number of points stays (relatively) low compared to
the total inserted.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Enable 100_000 benchmark

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Make our benchmark jump around

We need a known jumping around distribution of data. Sequential
insertion is a lie.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Avoid an extra loop to compute 'r'

This impacts sequential insertion but that's assumed to be rather
more rare than pseudo-random.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Fasted vec version so far

This is... this is about as far as I can get the Vec, I think, without
some major brainstorm.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Begin moving to a Batch style

This commit commit moves our implementation to the Batch style
discussed in the original CKMS paper. We are now trading off
write time for query time. This breaks the existing API and requires
an explicit flush before query.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Introduce selist

This is an experiment that I feel is already a dead-end. The
original paper discusses keeping their points in a linked list
which is... too damn slow. So here we are with a less slow linked
list.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Check in selist

This is a naive implementation of an unrolled linked list, plus
benchmarks. It's not... great, compared to a Vec. I'm not sure if
we can improve on this. plflib's benchmarks[1] indicate that list
is not much better than vector, in terms of insertion performance
which is our main concern here.

[1] http://www.plflib.org/benchmarks_haswell_gcc.htm#rawperformance

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Improve selist search performance

This is a x24 speedup on selist::u16::primed_insertion. That's...
pretty fancy. It puts us ahead of baseline Vec performance, 23k
nanoseconds versus 48k nanoseconds.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Tidy up SEList a bit

I fiddled with a cursor for a while but didn't manage to get
good implementation going. Which, I bet we can improve the
insertion performance with a cursor but I'm going to let that
hang for a while.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Further tidy-up work

This set of changes incorporates clippy and rustfmt suggestions.
There's still TODOs to be done but it's closer to something worth
bundling up.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Incorporate SEList

It looks like this is going to be pretty slow compared to a Vec,
even if the 'primed' insertion is suitable. I think I'll drop this
for now and see if I can't think of something else. Even with
failing tests the benchmarks are... not great.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Only make util available at test time, update deps

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Final wrap-up

This has been a long chain of commits. There's interesting work in
here -- I think I could probably contribute SEList to intrusive
collections -- but nothing that scooted up the performance for Cernan.
The prometheus sink in cernan would need to have some kind of quantile
summary that doesn't keep a sub-linear but growing number of points.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Cargo fmt etc etc

Signed-off-by: Brian L. Troutwine <blt@postmates.com>

* Remove flush, improve documentation

Signed-off-by: Brian L. Troutwine <blt@postmates.com>