Bookmark commit
In this commit we're set to run comparative benchmarks for Vec and
Store. The results were not promising, with Vec mean being 493ms
and Store mean being 546. Better SD and MAD on Vec too.

The computation of 'r' can be stored at the top of each inner store
and we can keep a sorted buffer for insertion. I guess the thing to
do is benchmark Vec against Store, see where we end up.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>