Go all-in on Store, perform compaction of inner data
This commit moves CKMS fully onto store with the understanding
that the current implementation is pokier than just plain Vec on
account of iteration on insertion. The expectation is that we
can resolve this.

Benchmarking -- not shown -- indicates that Store can do inserts
about twice as fast as Vec _if_ we don't have to seek over the
whole thing before inserting.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>