Better footing for benchmarks
I've removed sum from the interface of CKMS because if the type
is especially small I don't have a wrapping_add available to me.
Oops. Up to the client to keep track of this, as a result.

Well, maybe if I look into Wrapping?

Anyway! We're now on a good footing to benchmark this thing.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>