Simplfy the test further.
I think the issue that was pointed in #27 is an error with the user
supplied quantile code. Note that `error_nominal_test` and
`error_nominal_with_merge_test` assert that error bounds are maintained
without triggering the reported failure. This commit simplifies the
supplied test code and indicates that the disagreement happens with the
user quantile calculation. Note that the calculated rank and user index
are quite different.